+++ pwm.h
@@ -5,4 +5,4 @@ void Init_Timers(void);
 void Timer_RCC_Configuration(void);
 void Timer_GPIO_Enable(void);
 void Timer_GPIO_Disable(void);
-int PWM_Set(int Channel, int Pulse);
+int8_t PWM_Set(uint8_t Channel, uint16_t Pulse);
